streaming.py
import os
from dotenv import load_dotenv
from agents import Agent , AsyncOpenAI, OpenAIChatCompletionsModel, Runner ,  function_tool, handoff
from agents.run import RunConfig
from pydantic import BaseModel
from agents.extensions import handoff_filters

load_dotenv()

gemini_api_key = os.getenv("GEMINI_API_KEY")
if not gemini_api_key:
    raise ValueError("GEMINI_API_KEY is not set. Please ensure it is defined in your .env file.")

external_client = AsyncOpenAI(
    api_key = gemini_api_key,
    base_url="https://generativelanguage.googleapis.com/v1beta/openai/"
)

model = OpenAIChatCompletionsModel(
    model = 'gemini-2.0-flash',
    openai_client=external_client
)

config = RunConfig(
    model = model,
    model_provider = external_client,
    tracing_disabled = True
)

import asyncio
import random
from agents import Agent, ItemHelpers, Runner, function_tool

@function_tool
def how_many_jokes() -> int:
    return random.randint(1, 5)

async def main():
    agent = Agent(
        name = 'Joke Agent',
        instructions="First call the `how_many_jokes` tool, then tell that many jokes.",
        model = model,
        tools=[how_many_jokes],
    )  

    result =  Runner.run_streamed(
        agent,
        input= 'hello',
        run_config= config
    )

    print("=== Run starting ===")

    async for event in result.stream_events():
       # We'll ignore the raw responses event deltas
        if event.type == "raw_response_event":
            continue
        # When the agent updates, print that
        elif event.type == "agent_updated_stream_event":
            print(f"Agent updated: {event.new_agent.name}")
            continue
        # When items are generated, print them
        elif event.type == "run_item_stream_event":
            if event.item.type == "tool_call_item":
                print("-- Tool was called")
            elif event.item.type == "tool_call_output_item":
                print(f"-- Tool output: {event.item.output}")
            elif event.item.type == "message_output_item":
                print(f"-- Message output:\n {ItemHelpers.text_message_output(event.item)}")
            else:
                pass  # Ignore other event types

    print("=== Run complete ===")


if __name__ == "__main__":
    asyncio.run(main())